Use mdx-runtime on /ecosystem
