+++ pages/ecosystem.js
@@ -1,8 +1,9 @@
 import React from 'react'
+import MDX from '@mdx-js/runtime'
+import components from '../utils/mdx-components'
 
 import DocsLayout from '../components/DocsLayout'
 import { getReadme } from '../utils/githubApi'
-import md from '../components/md'
 import escape from '../utils/escape'
 import Loading from '../components/Loading'
 import Link from '../components/Link'
@@ -17,15 +18,16 @@ const Ecosystem = ({ readme, sidebarPages }) => (
     <p>
       This is an incomplete list of awesome things built with styled-components. If you have something to share, please add it to the <Link href="https://github.com/styled-components/awesome-styled-components" inline>awesome-styled-components</Link> repo on GitHub and it will automatically show up here!
     </p>
-    {typeof readme !== 'string' ? <Loading /> : md(
-        `
+    {typeof readme !== 'string' ? <Loading /> : (
+      <MDX components={components}>{`
         ${readme}
 
 ### Contribute
 
-If you know any projects build with styled components contributions and suggestions are always welcome !
+If you know any projects build with styled components contributions and suggestions are always welcome!
 Please read the [contribution guidelines](https://github.com/styled-components/awesome-styled-components/blob/master/contributing.md) first and submit a PR.
-        `)}
+      `}</MDX>
+    )}
   </DocsLayout>
 )
